# Top-N: Improve performance with large heaps, and correctly call Reduce